+++ Scripts/MMOGame/Networking/Map/MMOServerCashShopMessageHandlers.cs
@@ -146,7 +146,7 @@ namespace MultiplayerARPG.MMO
                             item = itemAmount.item,
                             level = 1,
                             amount = itemAmount.amount,
-                            randomSeed = (short)Random.Range(short.MinValue, short.MaxValue),
+                            randomSeed = Random.Range(int.MinValue, int.MaxValue),
                         });
                     }
                 }
